+++ airbyte-cdk/python/airbyte_cdk/sources/configurable/__init__.py
