+++ airbyte-cdk/python/airbyte_cdk/sources/configurable/checks/connection_checker.py
