+++ airbyte-cdk/python/airbyte_cdk/sources/configurable/requesters/paginators/paginator.py
