+++ airbyte-cdk/python/airbyte_cdk/sources/configurable/stream_slicers/stream_slicer.py
