+++ airbyte-cdk/python/unit_tests/sources/configurable/interpolation/__init__.py
